refactor: js -> ts로 확장자 변경 및 메인파일 src 폴더 내부로 이동
