+++ src/middleware/auth-optional-middleware.ts
