+++ src/middleware/image-middleware.ts
