+++ src/route/image-router.ts
